Revised arguments

+++ src/main/clone_scraper.sh
@@ -3,14 +3,16 @@
 # Listed below are the headers to be copied to the csv file when needed
 # Number of Files,Max Depth of Files,Number of Contributors,Number of Commits,Number of Merge Commits,Number of Branches,Number of Tags,Number of Links,Has README,Has SECURITY,Has CODE_OF_CONDUCT,Has CONTRIBUTING,Has ISSUE_TEMPLATE,Has PULL_REQUEST_TEMPLATE
 
-if [[ $# -ne 2 ]]; then
-  echo "usage: ./clone_scraper [fullpath_to_import_file] [fullpath_to_export_file]"
+if [[ $# -ne 4 ]]; then
+  echo "usage: ./clone_scraper [fullpath_to_import_file] [fullpath_to_export_file] [clone_repo_directory] [remove_option]"
   exit 
 fi
 
 # Store command line argument variables
 import_file=$1
 export_file=$2
+clone_repo_directory=$3
+remove_option=$4
 
 DATE=$(date)
 repo_list=()
@@ -24,6 +26,7 @@ done < "$import_file"
 
 function scrape {
   local arg=$1
+  local remove=$(echo "$2" | tr '[:upper:]' '[:lower:]')
   echo "$arg"
   git clone "$arg"
 
@@ -127,10 +130,13 @@ function scrape {
   # Output data to the csv file
   flock -x "$export_file" sh -c "echo '$arg,$num_files,$depth,$num_contributors,$num_commits,$num_merges,$num_branches,$num_tags,$num_links,$README,$SECURITY,$CONDUCT,$CONTRIBUTING,$ISSUE_TEMPLATE,$PULL_TEMPLATE' >> $export_file"
 
-  # CODE LEFT HERE JUST IN CASE: 
+  if [[ "$remove" == "yes" ]]; then
+    echo -e "y\ny\n" | rm -r $directory
+  fi
+
+# CODE LEFT HERE JUST IN CASE: 
   # Remove the cloned repository's directory
   # cd ..
-  echo -e "y\ny\n" | rm -r $directory
 
   # Compress the cloned repository
   # -c creates an archive, -z tells tar to use gzip, -f specifies file name of compressed file
@@ -146,6 +152,9 @@ function scrape {
 max_processes=10
 current_processes=0
 
+# Change into directory where cloned repos will be stored
+cd "$clone_repo_directory"
+
 for arg in "${repo_list[@]}"; do
       # Check if the maximum number of concurrent processes is reached
     if (( current_processes >= max_processes )); then
@@ -154,7 +163,7 @@ for arg in "${repo_list[@]}"; do
     fi
 
     # Run the function in the background
-    scrape "$arg" &
+    scrape "$arg" "$remove_option" &
     ((current_processes++))
 done
